refactor: move message reply logic to utils/replyDelete.ts
